ov002: promote dBgActor_c into a single C++ TU
The base of the level-object family becomes one production source,
src/actors/dBgActor_c.cpp, absorbing the eleven one-function shards that
used to carry it. The ROM build takes all eleven functions from a single
object, the way the cartridge's own build did.

WHAT THE CARTRIDGE PROVES ABOUT THE CLASS

  _ZTS  ov002 0x02108a04  "10dBgActor_c"
  _ZTI  ov002 0x021089ec  __si_class_type_info; +8 -> _ZTI8dActor_c
                          (arm9 0x0208e390), so the DIRECT base is
                          dActor_c and nothing else
  _ZTV  ov002 0x0210ae38  32 slots against dActor_c's 31
  size  0x320             the header's assert, corroborated by
                          daObjMc_Metalnet_c's factory literal (72c6dcfb6)

THE VTABLE was diffed slot by slot against _ZTV8dActor_c (arm9 0x0208e3a4).
Exactly three slots point inside ov002 -- the 16/17 destructor pair and slot
31, Kill, the one virtual this class adds. Every other slot still holds
dActor_c's own word, so nothing else is overridden.

THE DESTRUCTOR STAYS INLINE IN THE HEADER and Kill is defined here, which
makes Kill the key function: this TU emits _ZTV/_ZTI/_ZTS and both destructor
variants. The two shards for D1 and D0 each carried a forcing scaffold
(`p->~dBgActor_c()` and `delete p`) to make the compiler emit a variant they
did not own; this file needs neither, because owning the key function drags
both variants in. None of that data lies inside the licensed .text range
0x020ee42c..0x020eea84, so dsd supplies the cartridge's bytes and production
isolation discards the emitted duplicates. The manifest licenses all eleven
compiler-only outputs at their measured ROM homes -- this class's own trio
plus the inherited dActor_c/dBase_c/fBase_c chain, Vector3's inline
destructor, and the C1 the ROM never carried.

TWO SHARDS COULD NOT BE MERGED VERBATIM. UpdateKillByMegaChar carried its own
stand-in `dBgActor_c`, `dActor_c`, `Matrix4x3` and `Fix12`, which cannot
coexist with include/dBgActor_c.h in one TU. Reconciled against the real
header, the placeholder-vtable call became a plain virtual Kill() and the
cast-through-offset UpdatePos became a member call -- both free. The four
s16 angle accesses at 0x8c/0x8e/0x90/0x94 are NOT free: named-member spelling
lets mwcc CSE the field address and the function comes out 0x148 against the
ROM's 0x158. The raw-offset spelling is kept for those four, and only those
four, because the byte gate decides it.

func_ov002_020ee5d0 stays unnamed. It has one inbound relocation, from
UpdateKillByMegaChar, and no evidence of its original spelling.

PROOF

  tools/tubuild.py verify ov002/dBgActor_c
    11/11 MATCH, objisolate clean, reloc-destinations clean -> TEXT-VERIFIED
  tools/tubuild.py linkcheck ov002/dBgActor_c
    SCRATCH-LINK-VERIFIED; ROM sha256 identical to the baseline control
  tools/rombuild.py -j16 --no-rom
    106/106 exact, 100.000000% of compared bytes, mismatching 0
    source-built functions 11,088, reproducing 11,088
  static gates: check_src_tu_compiles, rename-ledger, profile-campaign,
    langmode ratchet, header-offsets, duplicate-sources, dead-references,
    tubuild-conflicts, port_refcheck, cpp_tu_compat --require-ready all clean;
    converted-baseline re-banked with tiers_ratchet --update (2597 -> 2593,
    the four removals logged as the consolidation they are)

Also adds notes/data/class-facts/dBgActor_c.json, the scout stage's ROM
evidence for this class.
