+++ config/converted-backslide-exceptions.jsonl
@@ -264,3 +264,7 @@
 {"path": "src/_ZN8ShipWing6RenderEv.cpp", "reason": "ov036/daObjRc_Hane_c promoted to a single intact TU (src/actors/d_a_obj_rc_hane.cpp); its seven legacy one-function sources are consolidated, so their individual CONVERTED rows retire into the TU's."}
 {"path": "src/_ZN8ShipWingD0Ev.cpp", "reason": "ov036/daObjRc_Hane_c promoted to a single intact TU (src/actors/d_a_obj_rc_hane.cpp); its seven legacy one-function sources are consolidated, so their individual CONVERTED rows retire into the TU's."}
 {"path": "src/_ZN8ShipWingD1Ev.cpp", "reason": "ov036/daObjRc_Hane_c promoted to a single intact TU (src/actors/d_a_obj_rc_hane.cpp); its seven legacy one-function sources are consolidated, so their individual CONVERTED rows retire into the TU's."}
+{"path": "src/_ZN10dBgActor_c21UpdateModelPosAndRotYEv.cpp", "reason": "ov002/dBgActor_c TU promotion: the four CONVERTED one-function sources are folded into src/actors/dBgActor_c.cpp, which is one physical entry. No function lost its tier; a many-to-one consolidation cannot keep four baseline paths."}
+{"path": "src/_ZN10dBgActor_cC2Ev.cpp", "reason": "ov002/dBgActor_c TU promotion: the four CONVERTED one-function sources are folded into src/actors/dBgActor_c.cpp, which is one physical entry. No function lost its tier; a many-to-one consolidation cannot keep four baseline paths."}
+{"path": "src/_ZN10dBgActor_cD0Ev.cpp", "reason": "ov002/dBgActor_c TU promotion: the four CONVERTED one-function sources are folded into src/actors/dBgActor_c.cpp, which is one physical entry. No function lost its tier; a many-to-one consolidation cannot keep four baseline paths."}
+{"path": "src/_ZN10dBgActor_cD1Ev.cpp", "reason": "ov002/dBgActor_c TU promotion: the four CONVERTED one-function sources are folded into src/actors/dBgActor_c.cpp, which is one physical entry. No function lost its tier; a many-to-one consolidation cannot keep four baseline paths."}
