+++ config/converted-baseline.json
@@ -7,7 +7,7 @@
     "no_codegen_trick",
     "no_mangled_refs"
   ],
-  "count": 2597,
+  "count": 2593,
   "converted": [
     "src/ARMMathLoadState.c",
     "src/ARMMathSaveState.c",
@@ -342,10 +342,6 @@
     "src/_ZN10StarSwitch6RenderEv.cpp",
     "src/_ZN10StarSwitchD0Ev.cpp",
     "src/_ZN10StarSwitchD1Ev.cpp",
-    "src/_ZN10dBgActor_c21UpdateModelPosAndRotYEv.cpp",
-    "src/_ZN10dBgActor_cC2Ev.cpp",
-    "src/_ZN10dBgActor_cD0Ev.cpp",
-    "src/_ZN10dBgActor_cD1Ev.cpp",
     "src/_ZN10dBgCh_Actr13SetGroundFlagEv.cpp",
     "src/_ZN10dBgCh_Actr13SetLimMovFlagEv.cpp",
     "src/_ZN10dBgCh_Actr15ClearGroundFlagEv.cpp",
@@ -2357,6 +2353,10 @@
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_c16CleanupResourcesEv",
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_cD0Ev",
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_cD1Ev",
+    "src/actors/d_a_obj_km3_kuruma.cpp#_ZN17daObjKm3_Kuruma_c13InitResourcesEv",
+    "src/actors/d_a_obj_km3_kuruma.cpp#_ZN17daObjKm3_Kuruma_c16CleanupResourcesEv",
+    "src/actors/d_a_obj_km3_kuruma.cpp#_ZN17daObjKm3_Kuruma_cD0Ev",
+    "src/actors/d_a_obj_km3_kuruma.cpp#_ZN17daObjKm3_Kuruma_cD1Ev",
     "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_c16CleanupResourcesEv",
     "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_c6RenderEv",
     "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_cD0Ev",
@@ -2552,10 +2552,6 @@
     "src/actors/daObjKm3_Kaitendai_c.cpp#_ZN20daObjKm3_Kaitendai_c16CleanupResourcesEv",
     "src/actors/daObjKm3_Kaitendai_c.cpp#_ZN20daObjKm3_Kaitendai_cD0Ev",
     "src/actors/daObjKm3_Kaitendai_c.cpp#_ZN20daObjKm3_Kaitendai_cD1Ev",
-    "src/actors/d_a_obj_km3_kuruma.cpp#_ZN17daObjKm3_Kuruma_c13InitResourcesEv",
-    "src/actors/d_a_obj_km3_kuruma.cpp#_ZN17daObjKm3_Kuruma_c16CleanupResourcesEv",
-    "src/actors/d_a_obj_km3_kuruma.cpp#_ZN17daObjKm3_Kuruma_cD0Ev",
-    "src/actors/d_a_obj_km3_kuruma.cpp#_ZN17daObjKm3_Kuruma_cD1Ev",
     "src/actors/daObjKm3_Kurumajiku_c.cpp#_ZN21daObjKm3_Kurumajiku_c13InitResourcesEv",
     "src/actors/daObjKm3_Kurumajiku_c.cpp#_ZN21daObjKm3_Kurumajiku_c16CleanupResourcesEv",
     "src/actors/daObjKm3_Kurumajiku_c.cpp#_ZN21daObjKm3_Kurumajiku_cD0Ev",
