+++ port/slice_gate9.txt
@@ -7,7 +7,10 @@ src/_ZN14ArrowSignRight13InitResourcesEv.cpp
 src/_ZN14ArrowSignRight16CleanupResourcesEv.cpp
 src/_ZN14ArrowSignRight8BehaviorEv.cpp
 src/_ZN14ArrowSignRight6RenderEv.cpp
-src/_ZN10dBgActor_cC2Ev.cpp
+# dBgActor_c is one production translation unit; the constructor,
+# UpdateClsnPosAndRot, UpdateModelPosAndRotY and IsClsnInRangeOnScreen
+# this slice used to name separately all live in it.
+src/actors/dBgActor_c.cpp
 src/_ZN8dActor_cC2Ev.cpp
 # SceneNode's constructor and Reset now share their production translation unit.
 src/actors/ActorBase_SceneNode.cpp
@@ -20,9 +23,6 @@ src/_ZN4dBgW7DisableEv.cpp
 src/_ZN4dBgW9IsEnabledEv.cpp
 src/_ZN10dBgW_KcMbg7SetFileEP8KCL_FileRK9Matrix4x35Fix12IiEsR10CLPS_Block.c
 src/_ZN8dActor_c9SetRangesE5Fix12IiES1_S1_S1_.cpp
-src/_ZN10dBgActor_c19UpdateClsnPosAndRotEv.cpp
-src/_ZN10dBgActor_c21IsClsnInRangeOnScreenE5Fix12IiES1_.cpp
-src/_ZN10dBgActor_c21UpdateModelPosAndRotYEv.cpp
 src/_ZN9ModelBase7SetFileEP8BMD_Fileii.cpp
 src/func_02039394.c
 src/func_020393a4.c
